+++ @navikt/core/react/src/layout/box/Box.tsx
@@ -3,21 +3,31 @@ import React, { forwardRef } from "react";
 import { OverridableComponent } from "../../util/types";
 import { getResponsiveProps } from "../utilities/css";
 import {
-  BackgroundToken,
+  BackgroundColorToken,
   BorderColorToken,
   BorderRadiiToken,
   ResponsiveProp,
   ShadowToken,
   SpaceDelimitedAttribute,
   SpacingScale,
+  SurfaceColorToken,
 } from "../utilities/types";
 
 export interface BoxProps extends React.HTMLAttributes<HTMLDivElement> {
-  /** CSS `background-color` property. Accepts a color token. */
-  background?: BackgroundToken;
-  /** CSS `border-color` property. Accepts a color token. */
+  /**
+   * CSS `background-color` property.
+   * Accepts a [background/surface color token](https://aksel.nav.no/grunnleggende/styling/design-tokens#afff774dad80).
+   */
+  background?: BackgroundColorToken | SurfaceColorToken;
+  /**
+   * CSS `border-color` property.
+   * Accepts a [border color token](https://aksel.nav.no/grunnleggende/styling/design-tokens#adb1767e2f87).
+   */
   borderColor?: BorderColorToken;
-  /** CSS `border-radius` property. Accepts a radius token, or an object of radius tokens for different breakpoints.
+  /**
+   * CSS `border-radius` property.
+   * Accepts a [radius token](https://aksel.nav.no/grunnleggende/styling/design-tokens#6d79c5605d31)
+   * or an object of radius tokens for different breakpoints.
    * @example
    * borderRadius='full'
    * borderRadius='0 full large small'
@@ -31,13 +41,19 @@ export interface BoxProps extends React.HTMLAttributes<HTMLDivElement> {
    * borderWidth='1 2 3 4'
    */
   borderWidth?: SpaceDelimitedAttribute<"0" | "1" | "2" | "3" | "4" | "5">;
-  /** Padding around children. Accepts a spacing token or an object of spacing tokens for different breakpoints.
+  /**
+   * Padding around children.
+   * Accepts a [spacing token](https://aksel.nav.no/grunnleggende/styling/design-tokens#0cc9fb32f213)
+   * or an object of spacing tokens for different breakpoints.
    * @example
    * padding='4'
    * padding={{xs: '2', sm: '3', md: '4', lg: '5', xl: '6'}}
    */
   padding?: ResponsiveProp<SpacingScale>;
-  /** Horizontal padding around children. Accepts a spacing token or an object of spacing tokens for different breakpoints.
+  /**
+   * Horizontal padding around children.
+   * Accepts a [spacing token](https://aksel.nav.no/grunnleggende/styling/design-tokens#0cc9fb32f213)
+   * or an object of spacing tokens for different breakpoints.
    * @example
    * paddingInline='4'
    * paddingInline='4 5'
@@ -46,7 +62,10 @@ export interface BoxProps extends React.HTMLAttributes<HTMLDivElement> {
   paddingInline?: ResponsiveProp<
     SpacingScale | `${SpacingScale} ${SpacingScale}`
   >;
-  /** Vertical padding around children. Accepts a spacing token or an object of spacing tokens for different breakpoints.
+  /**
+   * Vertical padding around children.
+   * Accepts a [spacing token](https://aksel.nav.no/grunnleggende/styling/design-tokens#0cc9fb32f213)
+   * or an object of spacing tokens for different breakpoints.
    * @example
    * paddingBlock='4'
    * paddingBlock='4 5'