+++ aksel.nav.no/website/components/website-modules/search/utils/group-results.ts
@@ -7,10 +7,11 @@ export function createSearchResult(
 ) {
   const groupedHits: GroupedHitsT = result?.reduce((prev, cur) => {
     if (cur.item._type in prev) {
-      return { ...prev, [cur.item._type]: [...prev[cur.item._type], cur] };
+      prev[cur.item._type].push(cur);
     } else {
-      return { ...prev, [cur.item._type]: [cur] };
+      prev[cur.item._type] = [cur];
     }
+    return prev;
   }, {});
 
   const response: SearchResultsT = {